Move HostServiceHandlerConfig into hostbridge-grpc-handler.ts
